feat: mvvm split user info, create userinfoadmintools container component
