+++ test/unit-tests/components/viewmodels/right_panel/UserInfoAdminToolsContainerViewModel-test.tsx
@@ -1,0 +1,16 @@
+/*
+Copyright 2025 New Vector Ltd.
+SPDX-License-Identifier: AGPL-3.0-only OR GPL-3.0-only
+Please see LICENSE files in the repository root for full details.
+*/
+
+import { useUserInfoAdminToolsContainerViewModel } from "../../../../../src/components/viewmodels/right_panel/UserInfoAdminToolsContainerViewModel";
+
+jest.mock("../../../../../src/components/views/elements/RoomTopic");
+
+describe("UserInfoAdminToolsContainerViewModel", () => {
+    it("should return the correct view model", () => {
+        // const vm = useUserInfoAdminToolsContainerViewModel({room, member, powerLevels});
+        // expect(vm).toBeDefined();
+    });
+});
